Fix environment variable and tidy
